// src/app/interfaces/lancamento.ts
export interface CarDetails {
    id: string;
    nome: string;
    preco: number;
    alturaCacamba: number;
    alturaVeiculo: number;
    alturaSolo: number;
    capacidadeCarga: number;
    motor: string;
    potencia: number;
    volumeCacamba: number;
    roda: string;
    image: string;
    selected?: boolean;
}